Move ocr validation to separate package
