--- src/main/java/uk/gov/hmcts/reform/bulkscanccdeventhandler/model/FormType.java
@@ -1,6 +1,0 @@
-package uk.gov.hmcts.reform.bulkscanccdeventhandler.model;
-
-public enum FormType {
-    CONTACT,
-    PERSONAL
-}
