+++ src/main/java/uk/gov/hmcts/reform/bulkscanccdeventhandler/model/OcrData.java
@@ -1,5 +1,7 @@
 package uk.gov.hmcts.reform.bulkscanccdeventhandler.model;
 
+import uk.gov.hmcts.reform.bulkscanccdeventhandler.ocrvalidation.model.in.OcrDataField;
+
 import java.util.List;
 
 public class OcrData {
