+++ src/main/java/uk/gov/hmcts/reform/bulkscanccdeventhandler/ocrvalidation/model/FormType.java
@@ -1,0 +1,6 @@
+package uk.gov.hmcts.reform.bulkscanccdeventhandler.ocrvalidation.model;
+
+public enum FormType {
+    CONTACT,
+    PERSONAL
+}
