+++ src/main/java/uk/gov/hmcts/reform/bulkscanccdeventhandler/services/OcrFieldRetriever.java
@@ -3,8 +3,8 @@ package uk.gov.hmcts.reform.bulkscanccdeventhandler.services;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import uk.gov.hmcts.reform.bulkscanccdeventhandler.model.CcdCollectionElement;
 import uk.gov.hmcts.reform.bulkscanccdeventhandler.model.OcrData;
-import uk.gov.hmcts.reform.bulkscanccdeventhandler.model.OcrDataField;
 import uk.gov.hmcts.reform.bulkscanccdeventhandler.model.ResultOrErrors;
+import uk.gov.hmcts.reform.bulkscanccdeventhandler.ocrvalidation.model.in.OcrDataField;
 
 import java.util.List;
